pyproject.toml
[project]
name = "summarizepdfs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.119.0",
    "langchain>=1.0.0",
    "langgraph>=1.0.0",
    "nltk>=3.9.2",
    "numpy>=2.3.4",
    "openai>=2.5.0",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.7",
    "pydantic>=2",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.1.1",
    "rapidfuzz>=3.14.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
    "tiktoken>=0.12.0",
    "typer>=0.19.2",
    "unstructured[docx,local-inference,pdf,pptx]>=0.18.15",
    "uvicorn>=0.38.0",
    "wandb>=0.22.2",
    "weave>=0.52.8",
]
